Actualizacion libreria zk-6.5.4
Primeros cambios hechos para que concuerde con cambios hechos en
proyecto adempiere. No Estan basados en los commits de Idempiere

+++ zkwebui/WEB-INF/src/org/adempiere/webui/component/NumberBox.java
@@ -147,7 +147,7 @@ public class NumberBox extends Div
     public void setValue(Object value)
     {
     	if (value == null)
-    		decimalBox.setValue(null);
+    		 decimalBox.setValue((BigDecimal) null);
     	else if (value instanceof BigDecimal)
     		decimalBox.setValue((BigDecimal) value);
     	else if (value instanceof Number)
